\name{estSSModel}

\alias{estSSModel}

\title{
  estSSModel
}

\description{
}

\usage{checkURL(wURL)}
%- maybe also 'usage' for other objects documented here.
\arguments{
}

\details{
}

\value{
}

\references{
%% ~put references to the literature/web site here ~
}

\author{
%%  ~~who you are~~
}

\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{kalmanRun}}, \code{\link{kalmanPred}}, \code{\link{arimaSS}}
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
