% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCinitial.R
\name{HCinitial}
\alias{HCinitial}
\title{Two-Step Method for Parameter Estimation of the Heckman Model}
\usage{
HCinitial(selection, outcome, data = sys.frame(sys.parent()))
}
\arguments{
\item{selection}{A formula for the selection equation.}

\item{outcome}{A formula for the outcome equation.}

\item{data}{A data frame containing the variables.}
}
\value{
A numeric vector containing the estimated parameters of the classic Heckman model:
\itemize{
  \item Coefficients from the selection equation (Probit model),
  \item Coefficients from the outcome equation (excluding the IMR),
  \item Estimated standard deviation of the error term (\code{sigma}),
  \item Estimated correlation between error terms (\code{rho}).
}
}
\description{
Estimates the parameters of the classical Heckman model using the two-step method.
}
\details{
This method is especially useful for obtaining initial values for maximum likelihood estimation.
The first step fits a Probit model to the selection equation.
The second step fits a linear regression to the outcome equation, adjusting for selection bias using the inverse Mills ratio.
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HCinitial(selectEq, outcomeEq, data = MEPS2001)

}
