% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanBS.R
\name{HeckmanBS}
\alias{HeckmanBS}
\title{Heckman BS Model fit Function}
\usage{
HeckmanBS(selection, outcome, data = sys.frame(sys.parent()), start = NULL)
}
\arguments{
\item{selection}{Selection equation.}

\item{outcome}{Primary Regression Equation.}

\item{data}{Database.}

\item{start}{initial values.}
}
\description{
Estimates the parameters of the Heckman-BS model
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeBS <- ambexp ~ age + female + educ + blhisp + totchr + ins
HeckmanBS(selectEq, outcomeBS, data = MEPS2001)
}
