% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanCL.R
\name{HeckmanCL}
\alias{HeckmanCL}
\title{Classic Heckman Model fit Function}
\usage{
HeckmanCL(selection, outcome, data = sys.frame(sys.parent()), start = NULL)
}
\arguments{
\item{selection}{Selection equation.}

\item{outcome}{Primary Regression Equation.}

\item{data}{Database.}

\item{start}{initial values.}
}
\description{
Estimates the parameters of the classic Heckman model
via Maximum Likelihood method. The initial start is obtained
via the two-step method.
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmanCL(selectEq, outcomeEq, data = MEPS2001)
}
