% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmanSK.R
\name{HeckmanSK}
\alias{HeckmanSK}
\title{Normal Skew Model fit Function}
\usage{
HeckmanSK(
  selection,
  outcome,
  data = sys.frame(sys.parent()),
  lambda,
  start = NULL
)
}
\arguments{
\item{selection}{Selection equation.}

\item{outcome}{Primary Regression Equation.}

\item{data}{Database.}

\item{lambda}{Initial start for asymmetry parameter.}

\item{start}{initial values.}
}
\description{
Estimates the parameters of the Normal Skew model
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmanSK(selectEq, outcomeEq, data = MEPS2001, lambda = 1)
}
