% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeckmantS.R
\name{HeckmantS}
\alias{HeckmantS}
\title{Heckman-t Model fit Function}
\usage{
HeckmantS(selection, outcome, data = sys.frame(sys.parent()), df, start = NULL)
}
\arguments{
\item{selection}{Selection equation.}

\item{outcome}{Primary Regression Equation.}

\item{data}{Database.}

\item{df}{Initial start to the degree of freedom.}

\item{start}{initial values.}
}
\description{
Estimates the parameters of the Heckman-t model
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
HeckmantS(selectEq, outcomeEq, data = MEPS2001, df=12)
}
