% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR.R
\name{IMR}
\alias{IMR}
\title{Inverse Mills Ratio}
\usage{
IMR(selection, data = sys.frame(sys.parent()))
}
\arguments{
\item{selection}{A formula for the selection equation.}

\item{data}{A data frame containing the variables.}
}
\value{
A column vector (matrix with one column) containing the Inverse Mills Ratio
computed from the Probit model fitted to the selection equation.
}
\description{
Computes the column vector of the Inverse Mills Ratio (IMR)
from a Probit selection equation.
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
IMR(selectEq, data = MEPS2001)

}
