% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{step2}
\alias{step2}
\title{Heckman's two-step method}
\usage{
step2(YS, XS, YO, XO)
}
\arguments{
\item{YS}{Selection vector.}

\item{XS}{Selection Matrix.}

\item{YO}{Interest vector.}

\item{XO}{Matrix of the equation of interest.}
}
\description{
Estimate model parameters via two-step method
}
