% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostep.R
\name{twostep}
\alias{twostep}
\title{Heckman's two-step method}
\usage{
twostep(selection, outcome, data = sys.frame(sys.parent()))
}
\arguments{
\item{selection}{A formula for the selection equation.}

\item{outcome}{A formula for the outcome equation.}

\item{data}{A data frame containing the variables used in the model.}
}
\value{
Returns a numeric vector with the parameter estimates of the Classical Heckman model using the two-step method.
For more information, see \insertCite{heckman1979sample;textual}{ssmodels}
}
\description{
Estimate model parameters via two-step method.
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
twostep(selectEq, outcomeEq)

}
\references{
{
\insertAllCited{}
}
}
