% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_components.R
\name{extract_model_components}
\alias{extract_model_components}
\title{Extract model components from formulas}
\usage{
extract_model_components(
  selection,
  outcome,
  data,
  outcomeS = NULL,
  outcomeC = NULL,
  drop.levels = TRUE
)
}
\arguments{
\item{selection}{A formula representing the selection equation.}

\item{outcome}{A formula representing the outcome equation.}

\item{data}{A data frame containing all variables used in the formulas.}

\item{outcomeS}{Optional matrix or formula of covariates for the dispersion (sigma) parameter.}

\item{outcomeC}{Optional matrix or formula of covariates for the correlation (rho) parameter.}

\item{drop.levels}{Logical; if TRUE, unused factor levels are dropped.}
}
\value{
A list containing:
\itemize{
  \item \code{XS}: model matrix for selection equation.
  \item \code{YS}: response vector for selection equation.
  \item \code{NXS}: number of covariates in the selection model.
  \item \code{XO}: model matrix for outcome equation.
  \item \code{YO}: response vector for outcome equation.
  \item \code{NXO}: number of covariates in the outcome model.
  \item \code{Msigma}: matrix for modeling dispersion (NULL if not specified).
  \item \code{NE}: number of covariates for the dispersion model.
  \item \code{Mrho}: matrix for modeling correlation (NULL if not specified).
  \item \code{NV}: number of covariates for the correlation model.
  \item \code{YSLevels}: factor levels of the selection response variable.
}
}
\description{
This function extracts the model frame, model matrix, and model response from formulas and a dataset.
It is used in models such as HeckmanCL, HeckmanGe, HeckmanSK, HeckmanBS, and HeckmantS.
Optionally, it handles covariate matrices for modeling dispersion and correlation parameters.
}
