% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR.R
\name{IMR}
\alias{IMR}
\title{Inverse Mills Ratio}
\usage{
IMR(selection, data = sys.frame(sys.parent()))
}
\arguments{
\item{selection}{Selection equation.}

\item{data}{Database.}
}
\description{
Column vector of the inverse ratio of Mills
}
