% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMR.R
\name{IMR}
\alias{IMR}
\title{Inverse Mills Ratio}
\usage{
IMR(selection, data = sys.frame(sys.parent()))
}
\arguments{
\item{selection}{Selection equation.}

\item{data}{Database.}
}
\value{
Return column vector of the inverse ratio of Mills
}
\description{
Column vector of the inverse ratio of Mills
}
\examples{
data(MEPS2001)
attach(MEPS2001)
selectEq <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
IMR(selectEq)
}
