% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmanBS.R
\name{summary.HeckmanBS}
\alias{summary.HeckmanBS}
\title{Summary of Birnbaum-Saunders Heckman Model}
\usage{
\method{summary}{HeckmanBS}(object, ...)
}
\arguments{
\item{object}{HeckmanBS class object.}

\item{...}{others functions.}
}
\value{
Print estimates of the parameters of the Heckman-BS model
}
\description{
Summary of Birnbaum-Saunders Heckman Model
}
