% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmanCL.R
\name{summary.HeckmanCL}
\alias{summary.HeckmanCL}
\title{Summary of Classic Heckman Model}
\usage{
\method{summary}{HeckmanCL}(object, ...)
}
\arguments{
\item{object}{HeckmanCL class object.}

\item{...}{others functions.}
}
\value{
Print estimates of the parameters of the Classic Heckman model
}
\description{
Summary of Classic Heckman Model
}
