% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmanGe.R
\name{summary.HeckmanGe}
\alias{summary.HeckmanGe}
\title{Summary of Generalized Heckman Model}
\usage{
\method{summary}{HeckmanGe}(object, ...)
}
\arguments{
\item{object}{HeckmanGe class object.}

\item{...}{others functions.}
}
\value{
Print estimates of the parameters of the Generalized Heckman model
}
\description{
Summary of Generalized Heckman Model
}
