% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmanSK.R
\name{summary.HeckmanSK}
\alias{summary.HeckmanSK}
\title{Summary of Skew-Normal Heckman Model}
\usage{
\method{summary}{HeckmanSK}(object, ...)
}
\arguments{
\item{object}{An object of class \code{HeckmanSK}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Prints the summary output including coefficient tables and model fit statistics.
}
\description{
Print estimates of the parameters of the Heckman-SK model using Maximum Likelihood Estimation.
}
