% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmanSK.R
\name{summary.HeckmanSK}
\alias{summary.HeckmanSK}
\title{Summary of Skew-Normal Heckman Model}
\usage{
\method{summary}{HeckmanSK}(object, ...)
}
\arguments{
\item{object}{HeckmanSK class object.}

\item{...}{others functions.}
}
\value{
Print estimates of the parameters of the Heckman-SK model
}
\description{
Summary of Skew-Normal Heckman Model
}
