% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmantS.R
\name{summary.HeckmantS}
\alias{summary.HeckmantS}
\title{Summary of Heckman-ts Model}
\usage{
\method{summary}{HeckmantS}(object, ...)
}
\arguments{
\item{object}{HeckmantS class object.}

\item{...}{others functions.}
}
\description{
Print estimates of the parameters of the Heckman-ts model
}
