% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.HeckmantS.R
\name{summary.HeckmantS}
\alias{summary.HeckmantS}
\title{Summary of Heckman-t Model}
\usage{
\method{summary}{HeckmantS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{HeckmantS}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Prints the summary output including coefficient tables and model fit statistics.
}
\description{
Print estimates of the parameters of the Heckman-t model using Maximum Likelihood Estimation.
}
