% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostep.R
\name{twostep}
\alias{twostep}
\title{Heckman's two-step method}
\usage{
twostep(selection, outcome, data = sys.frame(sys.parent()))
}
\arguments{
\item{selection}{Selection equation.}

\item{outcome}{Primary Regression Equation.}

\item{data}{Database.}
}
\description{
Estimate model parameters via two-step method
}
